+++ NEWS.md
@@ -3,6 +3,9 @@
 * `@inheritDotParams` automatically ignores arguments that can't be inherited
   through `...` because they are used by the current function (@mjskay, #885).
 
+* Escaped `'` and `"` in strings in examples are no longer double escaped 
+  (#873).
+
 * The default formatting for function usage that spans multiple lines has
   now changed. Previously it was wrapped to produce the fewest number of
   lines, e.g.: